Updates Server.js with Yelp API Routes, Fixes Some Bugs
* Line 20 was throwing an error in my environment. Added a try/catch
  block so that if mongoose.connect throws an error `Error: Trying to
  open unclosed connection' the error is handled and an appropriate
  workaround is in place.
* Adds lines 132-157 to include Yelp API token and two POST routes
  * /api/yelpSearch
  * /api/yelpBusiness

With these changes signups are working, and trips can be created. With
schema changes, Activities can no longer be added to Trips. Instead,
we need a new form to add Itinerary items to Trips, and Activities to
Itinerary items.

Made a minor change to app.js because signup was throwing client-side
errors. This may create a minor merge conflict.