Create Server Configuration File and POST /api/locations Route
In this commit, we also comment out all calls to authentication in
routes for testing purposes.

Configuration variables have been moved to a config.js file.

A route for the Google Places API has been included. This expects an
object with a query property set equal to a string that contains a
location description.