Change Naming Convention of Server-side Models
Next we will change our database models to use the redesigned schema.
This will break the application. So in the meantime, we will create new
model schemas in the models/model.js files leaving the
models/userModel.js file intact so anyone working on client-side can
still work and test while the changes are being integrated.